Started converting to a library
Removed all of the code from the main sketch file. Created a library to
house code for DCC interation. Library currently creates a command object
and generates a collection of bytes that encode the command that should be
sent to the DCC decoder. Command must first be converted to a proper
string of DCC bits before being queued and sent.

Sketch file now contains some simple code to intialize the library and
send command parameters in human-readable form.